+++ testermint/src/test/kotlin/InvalidationTests.kt
@@ -1,7 +1,15 @@
 import com.productscience.*
+import com.productscience.data.AppState
+import com.productscience.data.BandwidthLimitsParams
+import com.productscience.data.Decimal
+import com.productscience.data.EpochParams
+import com.productscience.data.InferenceParams
 import com.productscience.data.InferencePayload
+import com.productscience.data.InferenceState
 import com.productscience.data.InferenceStatus
+import com.productscience.data.ValidationParams
 import com.productscience.data.getParticipant
+import com.productscience.data.spec
 import kotlinx.coroutines.asCoroutineDispatcher
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
@@ -11,6 +19,7 @@ import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Timeout
 import org.tinylog.kotlin.Logger
+import java.util.*
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
 import kotlin.test.assertNotNull
@@ -20,17 +29,23 @@ class InvalidationTests : TestermintTest() {
     @Timeout(15, unit = TimeUnit.MINUTES)
     @Order(Int.MAX_VALUE - 1)
     fun `test invalid gets removed and restored`() {
-        val (cluster, genesis) = initCluster(mergeSpec = alwaysValidate)
+        val (cluster, genesis) = initCluster(mergeSpec = alwaysValidate, reboot = true)
         cluster.allPairs.forEach { pair ->
             pair.waitForMlNodesToLoad()
         }
-        genesis.waitForNextInferenceWindow()
+        genesis.waitForStage(EpochStage.SET_NEW_VALIDATORS)
+        genesis.node.waitForNextBlock(3)
 
         val dispatcher = Executors.newFixedThreadPool(10).asCoroutineDispatcher()
         runBlocking(dispatcher) {
-            val deferreds = (1..10).map {
+            // Each parallel inference needs a unique request to avoid duplicate inferenceIds
+            // (timestamp-based IDs can collide when started at the same nanosecond)
+            // Inject a unique nonce field into the JSON request
+            val deferreds = (1..10).map { index ->
                 async {
-                    InferenceTestHelper(cluster, genesis, responsePayload = "Invalid JSON!!").runFullInference()
+                    val nonce = "$index-${UUID.randomUUID()}"
+                    val uniqueRequest = inferenceRequest.replaceFirst("}", ", \"_nonce\": \"$nonce\"}")
+                    InferenceTestHelper(cluster, genesis, request = uniqueRequest, responsePayload = "Invalid JSON!!").runFullInference()
                 }
             }
             deferreds.awaitAll()
@@ -40,7 +55,7 @@ class InvalidationTests : TestermintTest() {
 
         genesis.markNeedsReboot()
         logSection("Waiting for removal")
-        genesis.node.waitForNextBlock(2)
+        genesis.node.waitForNextBlock(5)
         val participants = genesis.api.getActiveParticipants()
         val excluded = participants.excludedParticipants.firstOrNull()
         assertNotNull(excluded, "Participant was not excluded")
@@ -48,7 +63,7 @@ class InvalidationTests : TestermintTest() {
         val genesisValidatorInfo = genesis.node.getValidatorInfo()
         val validators = genesis.node.getValidators()
         assertThat(validators.validators).hasSize(3)
-        val genesisValidator = validators.validators.first { it.consensusPubkey.value == genesisValidatorInfo.key }
+        val genesisValidator = validators.validators.first { it.consensusPubkey.value ==  genesisValidatorInfo.key }
         assertThat(genesisValidator.tokens).isEqualTo(0)
         genesis.waitForNextEpoch()
         val newParticipants = genesis.api.getActiveParticipants()
@@ -57,7 +72,6 @@ class InvalidationTests : TestermintTest() {
         assertNotNull(removedRestored, "Excluded participant was not restored")
     }
 
-    // 6.5m
     @Test
     fun `test valid with invalid validator gets validated`() {
         val (cluster, genesis) = initCluster(mergeSpec = alwaysValidate)
@@ -68,6 +82,8 @@ class InvalidationTests : TestermintTest() {
         val oddPair = cluster.joinPairs.last()
         oddPair.mock?.setInferenceResponse(defaultInferenceResponseObject.withMissingLogit())
         logSection("Getting invalid invalidation")
+        genesis.waitForStage(EpochStage.SET_NEW_VALIDATORS)
+        genesis.node.waitForNextBlock(3)
         val invalidResult =
             generateSequence { getInferenceResult(genesis) }
                 .first { it.executorBefore.id != oddPair.node.getColdAddress() }
@@ -87,13 +103,13 @@ class InvalidationTests : TestermintTest() {
     fun `test invalid gets marked invalid`() {
         var tries = 3
         val (cluster, genesis) = initCluster(reboot = true)
-        genesis.waitForNextInferenceWindow(10)
         val oddPair = cluster.joinPairs.last()
         val badResponse = defaultInferenceResponseObject.withMissingLogit()
         oddPair.mock?.setInferenceResponse(badResponse)
         var newState: InferencePayload
         do {
             logSection("Trying to get invalid inference. Tries left: $tries")
+            genesis.waitForNextInferenceWindow()
             newState = getInferenceValidationState(genesis, oddPair)
         } while (newState.statusEnum != InferenceStatus.INVALIDATED && tries-- > 0)
         logSection("Verifying invalidation")
@@ -119,5 +135,26 @@ class InvalidationTests : TestermintTest() {
         assertThat(inferencePayload.inference.status).isEqualTo(InferenceStatus.INVALIDATED.value)
     }
 
+    companion object {
+        val alwaysValidate = spec {
+            this[AppState::inference] = spec<InferenceState> {
+                this[InferenceState::params] = spec<InferenceParams> {
+                    this[InferenceParams::validationParams] = spec<ValidationParams> {
+                        this[ValidationParams::minValidationAverage] = Decimal.fromDouble(100.0)
+                        this[ValidationParams::maxValidationAverage] = Decimal.fromDouble(100.0)
+                        this[ValidationParams::downtimeHThreshold] = Decimal.fromDouble(100.0)
 
+                    }
+                    this[InferenceParams::bandwidthLimitsParams] = spec<BandwidthLimitsParams> {
+                        this[BandwidthLimitsParams::minimumConcurrentInvalidations] = 100L
+                    }
+                    this[InferenceParams::epochParams] = spec<EpochParams> {
+                        this[EpochParams::inferencePruningEpochThreshold] = 100L
+                        this[EpochParams::epochLength] = 20L
+                        this[EpochParams::epochShift] = 15
+                    }
+                }
+            }
+        }
+    }
 }
